+++ x/crosschain/keeper/keeper_cross_chain_tx.go
@@ -33,7 +33,7 @@ func (k Keeper) SetCctxAndNonceToCctxAndInTxHashToCctx(ctx sdk.Context, cctx typ
 	}
 	k.SetInTxHashToCctx(ctx, in)
 
-	tss, found := k.GetTSS(ctx)
+	tss, found := k.zetaObserverKeeper.GetTSS(ctx)
 	if !found {
 		return
 	}